change: make BitcoinScript.Parse() not panicing
Previously the BitcoinScript.ParseWithPanic() would panic on script
data pushes that would push-after-end-off-script (i.e. the OpCode
would specify a push of more bytes than remaining in the script).
Since that's quite common in coinbase transactions the parser would
panic regulary.

This solves this by not panicing on these pushes. The ParsedOpCodes
just contain what is left from the script.